feat: homepage buying 관련 수정
