+++ src/pages/main/HomePage.jsx
@@ -5,31 +5,20 @@ import TopBar from "../../components/_common/TopBar";
 import { Wrapper } from "../../components/_common/CommonExport";
 
 import HomeMyContent from "../../components/main/HomeMyContent";
-import HomeOthersContent from "../../components/main/HomeOthersContent";
 import HomeBuyingContent from "../../components/main/HomeBuyingContent";
 
 const HomePage = () => {
   const [myFilterActive, setMyFilterActive] = useState(true);
-  const [othersFilterActive, setOthersFilterActive] = useState(false);
   const [buyingFilterActive, setBuyingFilterActive] = useState(false);
-  const [paymentInfo, setPaymentInfo] = useState(false);
 
   const handleMyFilterClick = () => {
     setMyFilterActive(true);
-    setOthersFilterActive(false);
     setBuyingFilterActive(false);
   };
 
-  const handleOthersFilterClick = () => {
-    if (paymentInfo) {
-      setOthersFilterActive(true);
-      setMyFilterActive(false);
-      setBuyingFilterActive(false);
-    } else {
-      setBuyingFilterActive(true);
-      setMyFilterActive(false);
-      setOthersFilterActive(false);
-    }
+  const handleBuyingFilterClick = () => {
+    setMyFilterActive(false);
+    setBuyingFilterActive(true);
   };
 
   return (
@@ -46,16 +35,14 @@ const HomePage = () => {
             <span>MY</span>
           </MyFilter>
           {/*결제 정보에 따라 텍스트 다르게*/}
-          <OthersFilter
-            onClick={handleOthersFilterClick}
-            $active={buyingFilterActive || othersFilterActive}
+          <BuyingFilter
+            onClick={handleBuyingFilterClick}
+            $active={buyingFilterActive}
           >
-            <span>{paymentInfo ? "Others" : "Buying"}</span>
-          </OthersFilter>
+            <span>Buying</span>
+          </BuyingFilter>
         </Filters>
-
         {myFilterActive && <HomeMyContent />}
-        {othersFilterActive && <HomeOthersContent />}
         {buyingFilterActive && <HomeBuyingContent />}
       </Wrapper>
     </>
@@ -105,7 +92,7 @@ const MyFilter = styled(FilterStyle)`
   }
 `;
 
-const OthersFilter = styled(FilterStyle)`
+const BuyingFilter = styled(FilterStyle)`
   border-left: none;
   border-right: none;
   display: flex;
